release

+++ main.py
@@ -1,6 +1,6 @@
 import logging
 import os
-
+from message import inoagent_message
 import telebot
 from flask import Flask, request
 from telebot import TeleBot, types
@@ -14,9 +14,10 @@ logger.setLevel(logging.DEBUG)
 
 @bot.message_handler()
 def main(m):
-    x = m.reply_to_message.message_id if hasattr(m.reply_to_message, "message_id") else "None"
-    bot.send_message(m.chat.id, x)
-    bot.delete_message(m.chat.id, m.message_id)
+    if m.text.lower() == "иноагент" and hasattr(m.reply_to_message, "message_id"):
+        inoagent = m.reply_to_message.message_id
+        bot.delete_message(m.chat.id, m.message_id)
+        bot.reply_to(inoagent, inoagent_message)
 
 
 @server.route(f'/{BOT_TOKEN}', methods=['POST'])
